+++ pytsite/auth_log/__init__.py
@@ -15,6 +15,7 @@ def __init():
     odm.register_model('auth_log', _odm_model.AuthLog)
     events.listen('pytsite.auth.login', _eh.auth_login)
     events.listen('pytsite.auth.logout', _eh.auth_logout)
+    events.listen('pytsite.auth.login_error', _eh.auth_login_error)
 
     admin_href = router.ep_url('pytsite.odm_ui.ep.browse', {'model': 'auth_log'})
     admin.sidebar.add_menu('auth', 'auth_log', 'pytsite.auth_log@log', admin_href, 'fa fa-history',