+++ pytsite/auth_log/_api.py
@@ -1,0 +1,10 @@
+"""PytSite Auth Log API Functions.
+"""
+__author__ = 'Alexander Shepetko'
+__email__ = 'a@shepetko.com'
+__license__ = 'MIT'
+
+
+SEVERITY_ERROR = 3
+SEVERITY_WARNING = 4
+SEVERITY_INFO = 6